Move otherField.js to /utils
